requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dnspython==2.0.0
eventlet==0.25.2
Flask==1.1.2
Flask-SocketIO==4.3.0
greenlet==0.4.16
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.1.3
monotonic==1.5
numpy==1.18.1
pandas==1.0.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-engineio==3.13.0
python-socketio==4.5.1
pytz==2020.1
requests==2.23.0
six==1.14.0
tornado==6.0.4
urllib3==1.25.9
websocket-client==0.57.0
Werkzeug==1.0.1
wincertstore==0.2
jupyter==1.0.0
jupyterlab==3.3.2
scikit-learn==1.0.2